requirements.txt
# Web Scraping
beautifulsoup4==4.12.3
selenium==4.18.1
requests==2.31.0
lxml==5.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.27

# Data Processing
pandas==2.2.1
numpy==1.26.4
python-dateutil==2.9.0

# Machine Learning
scikit-learn==1.4.1.post1
xgboost==2.0.3

# Visualization
plotly==5.19.0
matplotlib==3.8.3
seaborn==0.13.2
folium==0.15.1

# Web Dashboard
streamlit==1.31.1

# NLP (for text analysis)
nltk==3.8.1

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
tqdm==4.66.2

# Testing
pytest==8.0.2
pytest-cov==4.1.0

# Code Quality
black==24.2.0
flake8==7.0.0
pylint==3.1.0
